requirements.txt
streamlit==1.30.0
tensorflow==2.14.0
numpy==1.25.0
pillow==10.0.0
